微修正
LikeButton が実装されている場所、コメントアウトしているコードの削除
